Fix chat input style issues (#224528)
Fix #211174
Fix microsoft/vscode-copilot#6083
