+++ src/vs/workbench/contrib/codeEditor/browser/simpleEditorOptions.ts
@@ -13,6 +13,9 @@ import { SelectionClipboardContributionID } from 'vs/workbench/contrib/codeEdito
 import { TabCompletionController } from 'vs/workbench/contrib/snippets/browser/tabCompletion';
 import { EditorExtensionsRegistry } from 'vs/editor/browser/editorExtensions';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { registerThemingParticipant } from 'vs/platform/theme/common/themeService';
+import { IDisposable } from 'vs/base/common/lifecycle';
+import { selectionBackground, inputBackground, inputForeground, editorSelectionBackground } from 'vs/platform/theme/common/colorRegistry';
 
 export function getSimpleEditorOptions(configurationService: IConfigurationService): IEditorOptions {
 	return {
@@ -59,4 +62,36 @@ export function getSimpleCodeEditorWidgetOptions(): ICodeEditorWidgetOptions {
 			TabCompletionController.ID,
 		])
 	};
+}
+
+/**
+ * Should be called to set the styling on editors that are appearing as just input boxes
+ * @param editorContainerSelector An element selector that will match the container of the editor
+ */
+export function setupSimpleEditorSelectionStyling(editorContainerSelector: string): IDisposable {
+	// Override styles in selections.ts
+	return registerThemingParticipant((theme, collector) => {
+		const selectionBackgroundColor = theme.getColor(selectionBackground);
+
+		if (selectionBackgroundColor) {
+			// Override inactive selection bg
+			const inputBackgroundColor = theme.getColor(inputBackground);
+			if (inputBackgroundColor) {
+				collector.addRule(`${editorContainerSelector} .monaco-editor-background { background-color: ${inputBackgroundColor}; } `);
+				collector.addRule(`${editorContainerSelector} .monaco-editor .selected-text { background-color: ${inputBackgroundColor.transparent(0.4)}; }`);
+			}
+
+			// Override selected fg
+			const inputForegroundColor = theme.getColor(inputForeground);
+			if (inputForegroundColor) {
+				collector.addRule(`${editorContainerSelector} .monaco-editor .view-line span.inline-selected-text { color: ${inputForegroundColor}; }`);
+			}
+
+			collector.addRule(`${editorContainerSelector} .monaco-editor .focused .selected-text { background-color: ${selectionBackgroundColor}; }`);
+		} else {
+			// Use editor selection color if theme has not set a selection background color
+			collector.addRule(`${editorContainerSelector} .monaco-editor .focused .selected-text { background-color: ${theme.getColor(editorSelectionBackground)}; }`);
+		}
+	});
+
 }
